+++ grid2op/Observation/BaseObservation.py
@@ -21,6 +21,8 @@ from scipy.sparse import csr_matrix
 
 # TODO fix "bug" when action not initalized, return nan in to_vect
 
+# TODO be consistent with gen_* and prod_* also in dictionaries
+
 
 class BaseObservation(GridObjects):
     """
@@ -51,14 +53,17 @@ class BaseObservation(GridObjects):
     day_of_week: ``int``
         The current day of the week (monday = 0 and sunday = 6)
 
-    prod_p: :class:`numpy.ndarray`, dtype:float
+    gen_p: :class:`numpy.ndarray`, dtype:float
         The active production value of each generator (expressed in MW).
+        (the old name "prod_p" is still usable)
 
-    prod_q: :class:`numpy.ndarray`, dtype:float
+    gen_q: :class:`numpy.ndarray`, dtype:float
         The reactive production value of each generator (expressed in MVar).
+        (the old name "prod_q" is still usable)
 
-    prod_v: :class:`numpy.ndarray`, dtype:float
+    gen_v: :class:`numpy.ndarray`, dtype:float
         The voltage magnitude of the bus to which each generator is connected (expressed in kV).
+        (the old name "prod_v" is still usable)
 
     load_p: :class:`numpy.ndarray`, dtype:float
         The active load value of each consumption (expressed in MW).
@@ -160,7 +165,7 @@ class BaseObservation(GridObjects):
     _attr_eq = ["line_status",
                 "topo_vect",
                 "timestep_overflow",
-                "prod_p", "prod_q", "prod_v",
+                "gen_p", "gen_q", "gen_v",
                 "load_p", "load_q", "load_v",
                 "p_or", "q_or", "v_or", "a_or",
                 "p_ex", "q_ex", "v_ex", "a_ex",
@@ -210,9 +215,9 @@ class BaseObservation(GridObjects):
         self.topo_vect = np.zeros(shape=self.dim_topo, dtype=dt_int)
 
         # generators information
-        self.prod_p = np.full(shape=self.n_gen, dtype=dt_float, fill_value=np.NaN)
-        self.prod_q = 1.0 * self.prod_p
-        self.prod_v = 1.0 * self.prod_p
+        self.gen_p = np.full(shape=self.n_gen, dtype=dt_float, fill_value=np.NaN)
+        self.gen_q = 1.0 * self.gen_p
+        self.gen_v = 1.0 * self.gen_p
         # loads information
         self.load_p = np.full(shape=self.n_load, dtype=dt_float, fill_value=np.NaN)
         self.load_q = 1.0 * self.load_p
@@ -245,8 +250,8 @@ class BaseObservation(GridObjects):
         self.day_of_week = dt_int(0)
 
         # redispatching
-        self.target_dispatch = 1.0 * self.prod_p
-        self.actual_dispatch = 1.0 * self.prod_p
+        self.target_dispatch = 1.0 * self.gen_p
+        self.actual_dispatch = 1.0 * self.gen_p
 
         # storage unit
         self.storage_charge = np.full(shape=self.n_storage, dtype=dt_float, fill_value=np.NaN)  # in MWh
@@ -257,6 +262,7 @@ class BaseObservation(GridObjects):
         self._connectivity_matrix_ = None
         self._bus_connectivity_matrix_ = None
         self._dictionnarized = None
+        self._vectorized = None
 
         # for shunt (these are not stored!)
         if self.shunts_data_available:
@@ -378,12 +384,12 @@ class BaseObservation(GridObjects):
         elif gen_id is not None:
             if line_id is not None or substation_id is not None or storage_id is not None:
                 raise Grid2OpException("You can only the inspect the effect of an action on one single element")
-            if gen_id >= len(self.prod_p):
+            if gen_id >= len(self.gen_p):
                 raise Grid2OpException("There are no generator of id \"gen_id={}\" in this grid.".format(gen_id))
 
-            res = {"p": self.prod_p[gen_id],
-                   "q": self.prod_q[gen_id],
-                   "v": self.prod_v[gen_id],
+            res = {"p": self.gen_p[gen_id],
+                   "q": self.gen_q[gen_id],
+                   "v": self.gen_v[gen_id],
                    "bus": self.topo_vect[self.gen_pos_topo_vect[gen_id]],
                    "sub_id": self.gen_to_subid[gen_id],
                    "target_dispatch": self.target_dispatch[gen_id],
@@ -470,9 +476,9 @@ class BaseObservation(GridObjects):
         self.topo_vect[:] = 0
 
         # generators information
-        self.prod_p[:] = np.NaN
-        self.prod_q[:] = np.NaN
-        self.prod_v[:] = np.NaN
+        self.gen_p[:] = np.NaN
+        self.gen_q[:] = np.NaN
+        self.gen_v[:] = np.NaN
         # loads information
         self.load_p[:] = np.NaN
         self.load_q[:] = np.NaN
@@ -537,9 +543,9 @@ class BaseObservation(GridObjects):
         - all prods are 0.
         - etc.
         """
-        self.prod_p[:] = 0.
-        self.prod_q[:] = 0.
-        self.prod_v[:] = 0.
+        self.gen_p[:] = 0.
+        self.gen_q[:] = 0.
+        self.gen_v[:] = 0.
         # loads information
         self.load_p[:] = 0.
         self.load_q[:] = 0.
@@ -1026,13 +1032,13 @@ class BaseObservation(GridObjects):
         lex_bus = tmplate[lex_bus]
         stor_bus = tmplate[stor_bus]
         if active_flow:
-            prod_vect = self.prod_p
+            prod_vect = self.gen_p
             load_vect = self.load_p
             or_vect = self.p_or
             ex_vect = self.p_ex
             stor_vect = self.storage_power
         else:
-            prod_vect = self.prod_q
+            prod_vect = self.gen_q
             load_vect = self.load_q
             or_vect = self.q_or
             ex_vect = self.q_ex
@@ -1100,9 +1106,9 @@ class BaseObservation(GridObjects):
 
         Returns
         -------
-        prod_p_f: ``numpy.ndarray``
+        gen_p_f: ``numpy.ndarray``
             The forecast generators active values
-        prod_v_f: ``numpy.ndarray``
+        gen_v_f: ``numpy.ndarray``
             The forecast generators voltage setpoins
         load_p_f: ``numpy.ndarray``
             The forecast load active consumption
@@ -1126,9 +1132,9 @@ class BaseObservation(GridObjects):
         if "load_q" in a["injection"]:
             load_q_f = a["injection"]["load_q"]
         tmp_arg = ~np.isfinite(prod_p_f)
-        prod_p_f[tmp_arg] = self.prod_p[tmp_arg]
+        prod_p_f[tmp_arg] = self.gen_p[tmp_arg]
         tmp_arg = ~np.isfinite(prod_v_f)
-        prod_v_f[tmp_arg] = self.prod_v[tmp_arg]
+        prod_v_f[tmp_arg] = self.gen_v[tmp_arg]
         tmp_arg = ~np.isfinite(load_p_f)
         load_p_f[tmp_arg] = self.load_p[tmp_arg]
         tmp_arg = ~np.isfinite(load_q_f)
@@ -1375,3 +1381,88 @@ class BaseObservation(GridObjects):
 
         """
         return self.topo_vect[self.storage_pos_topo_vect]
+
+    @property
+    def prod_p(self):
+        return self.gen_p
+
+    @property
+    def prod_q(self):
+        return self.gen_q
+
+    @property
+    def prod_v(self):
+        return self.gen_v
+
+    def _reset_matrices(self):
+        self._vectorized = None
+
+    def from_vect(self, vect, check_legit=True):
+        """
+        .. warning:: /!\\\\ Internal, do not use unless you know what you are doing /!\\\\
+
+            To reload an observation from a vector, use the "env.observation_space.from_vect()".
+
+        Convert back an observation represented as a vector into a proper observation.
+
+        Some conversion are done silently from float to the type of the corresponding observation attribute.
+
+        Parameters
+        ----------
+        vect: ``numpy.ndarray``
+            A representation of an BaseObservation in the form of a vector that is used to convert back the current
+            observation to be equal to the vect.
+
+        """
+
+        # reset the matrices
+        self._reset_matrices()
+        # and ensure everything is reloaded properly
+        super().from_vect(vect, check_legit=check_legit)
+
+    def to_dict(self):
+        """
+        Transform this observation as a dictionary. This dictionary allows you to inspect the state of this
+        observation and is simply a shortcut of the class instance.
+
+        Returns
+        -------
+        A dictionary representing the observation.
+
+        """
+        if self._dictionnarized is None:
+            self._dictionnarized = {}
+            self._dictionnarized["timestep_overflow"] = self.timestep_overflow
+            self._dictionnarized["line_status"] = self.line_status
+            self._dictionnarized["topo_vect"] = self.topo_vect
+            self._dictionnarized["loads"] = {}
+            self._dictionnarized["loads"]["p"] = self.load_p
+            self._dictionnarized["loads"]["q"] = self.load_q
+            self._dictionnarized["loads"]["v"] = self.load_v
+            self._dictionnarized["prods"] = {}
+            self._dictionnarized["prods"]["p"] = self.gen_p
+            self._dictionnarized["prods"]["q"] = self.gen_q
+            self._dictionnarized["prods"]["v"] = self.gen_v
+            self._dictionnarized["lines_or"] = {}
+            self._dictionnarized["lines_or"]["p"] = self.p_or
+            self._dictionnarized["lines_or"]["q"] = self.q_or
+            self._dictionnarized["lines_or"]["v"] = self.v_or
+            self._dictionnarized["lines_or"]["a"] = self.a_or
+            self._dictionnarized["lines_ex"] = {}
+            self._dictionnarized["lines_ex"]["p"] = self.p_ex
+            self._dictionnarized["lines_ex"]["q"] = self.q_ex
+            self._dictionnarized["lines_ex"]["v"] = self.v_ex
+            self._dictionnarized["lines_ex"]["a"] = self.a_ex
+            self._dictionnarized["rho"] = self.rho
+
+            self._dictionnarized["maintenance"] = {}
+            self._dictionnarized["maintenance"]['time_next_maintenance'] = self.time_next_maintenance
+            self._dictionnarized["maintenance"]['duration_next_maintenance'] = self.duration_next_maintenance
+            self._dictionnarized["cooldown"] = {}
+            self._dictionnarized["cooldown"]['line'] = self.time_before_cooldown_line
+            self._dictionnarized["cooldown"]['substation'] = self.time_before_cooldown_sub
+            self._dictionnarized["redispatching"] = {}
+            self._dictionnarized["redispatching"]["target_redispatch"] = self.target_dispatch
+            self._dictionnarized["redispatching"]["actual_dispatch"] = self.actual_dispatch
+
+        return self._dictionnarized