+++ grid2op/tests/test_Rules.py
@@ -123,8 +123,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
 
         self.helper_action.legal_action = RulesChecker(legalActClass=LookParam).legal_action
 
-        self.env.parameters.MAX_SUB_CHANGED = 2
-        self.env.parameters.MAX_LINE_STATUS_CHANGED = 2
+        self.env._parameters.MAX_SUB_CHANGED = 2
+        self.env._parameters.MAX_LINE_STATUS_CHANGED = 2
         _ = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                 "set_bus": {"substations_id": [(id_2, arr2)]},
                                 "change_line_status": arr_line1,
@@ -133,8 +133,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
                                check_legal=True)
 
         try:
-            self.env.parameters.MAX_SUB_CHANGED = 1
-            self.env.parameters.MAX_LINE_STATUS_CHANGED = 2
+            self.env._parameters.MAX_SUB_CHANGED = 1
+            self.env._parameters.MAX_LINE_STATUS_CHANGED = 2
             _ = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                     "set_bus": {"substations_id": [(id_2, arr2)]},
                                     "change_line_status": arr_line1,
@@ -146,8 +146,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
             pass
 
         try:
-            self.env.parameters.MAX_SUB_CHANGED = 2
-            self.env.parameters.MAX_LINE_STATUS_CHANGED = 1
+            self.env._parameters.MAX_SUB_CHANGED = 2
+            self.env._parameters.MAX_LINE_STATUS_CHANGED = 1
             _ = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                     "set_bus": {"substations_id": [(id_2, arr2)]},
                                     "change_line_status": arr_line1,
@@ -158,9 +158,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
         except IllegalAction:
             pass
 
-
-        self.env.parameters.MAX_SUB_CHANGED = 1
-        self.env.parameters.MAX_LINE_STATUS_CHANGED = 1
+        self.env._parameters.MAX_SUB_CHANGED = 1
+        self.env._parameters.MAX_LINE_STATUS_CHANGED = 1
         _ = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                 "set_line_status": arr_line2},
                                env=self.env,
@@ -180,8 +179,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
         arr_line2[id_line2] = 1
 
         self.helper_action.legal_action = RulesChecker(legalActClass=PreventReconnection).legal_action
-        self.env.parameters.MAX_SUB_CHANGED = 1
-        self.env.parameters.MAX_LINE_STATUS_CHANGED = 2
+        self.env._parameters.MAX_SUB_CHANGED = 1
+        self.env._parameters.MAX_LINE_STATUS_CHANGED = 2
         act = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                   "set_bus": {"substations_id": [(id_2, arr2)]},
                                   "change_line_status": arr_line1,
@@ -191,8 +190,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
         _ = self.env.step(act)
 
         try:
-            self.env.parameters.MAX_SUB_CHANGED = 2
-            self.env.parameters.MAX_LINE_STATUS_CHANGED = 1
+            self.env._parameters.MAX_SUB_CHANGED = 2
+            self.env._parameters.MAX_LINE_STATUS_CHANGED = 1
             self.env._times_before_line_status_actionable[id_line] = 1
             _ = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                     "set_bus": {"substations_id": [(id_2, arr2)]},
@@ -205,8 +204,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
             pass
 
         self.env._times_before_line_status_actionable[:] = 0
-        self.env.parameters.MAX_SUB_CHANGED = 2
-        self.env.parameters.MAX_LINE_STATUS_CHANGED = 1
+        self.env._parameters.MAX_SUB_CHANGED = 2
+        self.env._parameters.MAX_LINE_STATUS_CHANGED = 1
         self.env._times_before_line_status_actionable[1] = 1
         _ = self.helper_action({"change_bus": {"substations_id": [(id_1, arr1)]},
                                 "set_bus": {"substations_id": [(id_2, arr2)]},
@@ -234,8 +233,8 @@ class TestLoadingBackendFunc(unittest.TestCase):
 
         # i act a first time on powerline 15
         act = self.helper_action({"set_line_status": arr_line2},
-                               env=self.env,
-                               check_legal=True)
+                                 env=self.env,
+                                 check_legal=True)
         self.env.step(act)
         try:
             # i try to react on it, it should throw an IllegalAction exception.
@@ -291,7 +290,7 @@ class TestLoadingBackendFunc(unittest.TestCase):
         arr_line2[id_line2] = -1
 
         self.env._max_timestep_line_status_deactivated = 2
-        self.env.parameters.NB_TIMESTEP_LINE_STATUS_REMODIF = 2
+        self.env._parameters.NB_TIMESTEP_LINE_STATUS_REMODIF = 2
 
         self.helper_action.legal_action = RulesChecker(legalActClass=PreventReconnection).legal_action